Reuse configs for KotlinExtension and KotlinGradleExtension
